src/third_party/WebKit/LayoutTests/platform/win/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-flexbox-expected.txt

This is a testharness.js-based test.
FAIL Flex assert_equals: height expected "20px" but got "180px"
FAIL Inline flex assert_equals: height expected "20px" but got "180px"
Harness: the test ran to completion.
